@@ -21,20 +21,20 @@
     "heroku-postbuild": "ng build --configuration production"
   },
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^15.0.0",
-    "@angular/cli": "^17.3.17",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/localize": "^17.3.12",
+    "@angular/cli": "^18.2.19",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/localize": "^18.2.13",
     "@angular/material": "^15.0.0",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@angular/service-worker": "^17.3.12",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@angular/service-worker": "^18.2.13",
     "@ng-bootstrap/ng-bootstrap": "^15.1.2",
     "@ngrx/store": "^15.0.0",
     "@ngx-translate/core": "^15.0.0",
@@ -60,7 +60,7 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^20.0.0",
-    "@angular/language-service": "^17.3.12",
+    "@angular/language-service": "^18.2.13",
     "@babel/core": "^7.7.4",
     "@biesbjerg/ngx-translate-extract": "^7.0.4",
     "@fortawesome/fontawesome-free": "^5.8.1",
